Relocated calendar service and upcoming model.
