+++ homepage/src/app/app.module.ts
@@ -11,7 +11,7 @@ import { AppComponent } from './app.component';
 import { NavBarComponent } from './nav-bar/nav-bar.component';
 import { FooterComponent } from './footer/footer.component';
 import { UpcomingComponent } from './upcoming/upcoming.component';
-import { CalendarService } from './calendar.service';
+import { CalendarService } from './services/calendar.service';
 import { UserBubbleComponent } from './user-bubble/user-bubble.component';
 import { HomeComponent } from './home/home.component';
