--- homepage/src/app/calendar.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { CalendarService } from './calendar.service';
-
-describe('CalendarService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [CalendarService]
-    });
-  });
-
-  it('should be created', inject([CalendarService], (service: CalendarService) => {
-    expect(service).toBeTruthy();
-  }));
-});
